+++ src/utils/index.ts
@@ -1,5 +1,8 @@
 import { UnitConverter } from "./UnitConverter";
+import { LengthUnit, AngleUnit } from "./UnitConverter/units";
 
 export default {
   UnitConverter,
+  LengthUnit,
+  AngleUnit,
 };
